Add Items to DataTables (implement IEquatable for finding Items that are already in tables)
